fix rename bug
